Add Parser class to handle parsing of inputs
