+++ src/main/java/emiya/emiyaexception/EmiyaException.java
@@ -1,4 +1,4 @@
-package emiyaexception;
+package emiya.emiyaexception;
 
 public class EmiyaException extends Exception {
     public EmiyaException(String message) {
