+++ src/main/java/emiya/emiyaexception/EmptyDeadlineException.java
@@ -1,9 +1,9 @@
-package emiyaexception;
+package emiya.emiyaexception;
 
 public class EmptyDeadlineException extends EmiyaException{
     public EmptyDeadlineException() {
         super("-----------------------------------------\n" +
-                "Oh no! task.Deadline tasks cannot be empty! Please try again!\n"
+                "Oh no! Deadline tasks cannot be empty! Please try again!\n"
                 + "-----------------------------------------\n");
     }
 }
