+++ src/main/java/emiya/emiyaexception/EmptyEventException.java
@@ -1,9 +1,9 @@
-package emiyaexception;
+package emiya.emiyaexception;
 
 public class EmptyEventException extends EmiyaException{
     public EmptyEventException() {
         super("-----------------------------------------\n" +
-                "Oh no! task.Event tasks cannot be empty! Please try again!\n"
+                "Oh no! Event tasks cannot be empty! Please try again!\n"
                 + "-----------------------------------------\n");
     }
 }
