+++ src/main/java/emiya/emiyaexception/EmptyTodoException.java
@@ -1,4 +1,4 @@
-package emiyaexception;
+package emiya.emiyaexception;
 
 public class EmptyTodoException extends EmiyaException{
     public EmptyTodoException() {
