+++ src/main/java/emiya/emiyaexception/ListEmptyException.java
@@ -1,4 +1,4 @@
-package emiyaexception;
+package emiya.emiyaexception;
 
 public class ListEmptyException extends EmiyaException{
     public ListEmptyException() {
