+++ src/main/java/emiya/emiyaexception/NoByException.java
@@ -1,4 +1,4 @@
-package emiyaexception;
+package emiya.emiyaexception;
 
 public class NoByException extends EmiyaException{
     public NoByException() {
