+++ src/main/java/emiya/task/Event.java
@@ -1,8 +1,8 @@
-package task;
+package emiya.task;
 
-import datehandler.DateHandler;
-import emiyaexception.InvalidDateException;
-import emiyaexception.WrongDateFormatException;
+import emiya.datehandler.DateHandler;
+import emiya.emiyaexception.InvalidDateException;
+import emiya.emiyaexception.WrongDateFormatException;
 
 import java.time.LocalDateTime;
