+++ src/main/java/emiya/task/Task.java
@@ -1,4 +1,4 @@
-package task;
+package emiya.task;
 
 abstract public class Task {
     protected boolean isCompleted;
